CV3/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.7.0
openai>=1.30.0
ollama>=0.2.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
gtts>=2.5.1
python-dotenv>=1.0.0
requests>=2.31.0
